+++ api/handler.go
@@ -4,8 +4,11 @@
 
 package api
 
-import "net/http"
-import "github.com/tsuru/ingress-router/ingress"
+import (
+	"net/http"
+
+	"github.com/tsuru/ingress-router/router"
+)
 
 type handler func(http.ResponseWriter, *http.Request) error
 
@@ -16,7 +19,7 @@ func (h handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 func handleError(err error, w http.ResponseWriter, r *http.Request) {
 	if err != nil {
-		if err == ingress.ErrIngressAlreadyExists {
+		if err == router.ErrIngressAlreadyExists {
 			http.Error(w, err.Error(), http.StatusConflict)
 			return
 		}
