+++ kubernetes/ingress.go
@@ -7,8 +7,7 @@ package kubernetes
 import (
 	"fmt"
 
-	"github.com/tsuru/ingress-router/ingress"
-
+	"github.com/tsuru/ingress-router/router"
 	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
@@ -50,7 +49,7 @@ func (k *IngressService) Create(appName string) error {
 	}
 	_, err = client.Create(&i)
 	if k8sErrors.IsAlreadyExists(err) {
-		return ingress.ErrIngressAlreadyExists
+		return router.ErrIngressAlreadyExists
 	}
 	return err
 }
